requirements_backend.txt
PyYAML==6.0
Flask
transformers==4.30.2
diffusers[torch]==0.17.1
omegaconf
